+++ derive/src/interface.rs
@@ -165,10 +165,6 @@ pub fn generate(interface_args: &args::Interface) -> GeneratorResult<TokenStream
                 method_name,
             )
         };
-        let ty = match syn::parse_str::<syn::Type>(&ty.value()) {
-            Ok(ty) => ty,
-            Err(_) => return Err(Error::new_spanned(ty, "Expect type").into()),
-        };
         let mut calls = Vec::new();
         let mut use_params = Vec::new();
         let mut decl_params = Vec::new();
@@ -212,10 +208,6 @@ pub fn generate(interface_args: &args::Interface) -> GeneratorResult<TokenStream
             let name = interface_args
                 .rename_args
                 .rename(name, RenameTarget::Argument);
-            let ty = match syn::parse_str::<syn::Type>(&ty.value()) {
-                Ok(ty) => ty,
-                Err(_) => return Err(Error::new_spanned(ty, "Expect type").into()),
-            };
             decl_params.push(quote! { #ident: #ty });
             use_params.push(quote! { #ident });
 
@@ -276,7 +268,7 @@ pub fn generate(interface_args: &args::Interface) -> GeneratorResult<TokenStream
             .unwrap_or_else(|| quote! {::std::option::Option::None});
         let deprecation = gen_deprecation(deprecation, &crate_name);
 
-        let oty = OutputType::parse(&ty)?;
+        let oty = OutputType::parse(ty)?;
         let ty = match oty {
             OutputType::Value(ty) => ty,
             OutputType::Result(_, ty) => ty,
@@ -319,6 +311,7 @@ pub fn generate(interface_args: &args::Interface) -> GeneratorResult<TokenStream
                 override_from: #override_from,
                 visible: #visible,
                 compute_complexity: ::std::option::Option::None,
+                directive_invocations: ::std::vec![],
             });
         });